+++ filament/backend/src/vulkan/VulkanBuffer.cpp
@@ -18,6 +18,8 @@
 
 #include <utils/Panic.h>
 
+using namespace bluevk;
+
 namespace filament {
 namespace backend {